+++ changes.md
@@ -1,3 +1,8 @@
+#### 2020.V.02.15
+- Fixed flat template
+- Added ole wrapper
+- Added AlienVault (No key required)
+
 #### 2020.V.02.14
 - Added msg wrapper